Support Numpy 2.0 and Bump Python Version (#120)

+++ mo_gymnasium/envs/lunar_lander/lunar_lander.py
@@ -72,7 +72,7 @@ class MOLunarLander(LunarLander):  # no need for EzPickle, it's already in Lunar
             )
 
         if self.continuous:
-            action = np.clip(action, -1, +1).astype(np.float32)
+            action = np.clip(action, -1, +1).astype(np.float64)
         else:
             assert self.action_space.contains(action), f"{action!r} ({type(action)}) invalid "
